logger added

+++ src/main/scala/mongo/MongoService.scala
@@ -4,9 +4,11 @@ import com.mongodb.client.model.Accumulators.*
 import com.mongodb.client.model.Aggregates.*
 import com.mongodb.client.model.Filters
 import com.mongodb.client.model.Sorts.*
+import org.apache.logging.log4j.LogManager
 import org.bson.Document
 import spray.json.*
 import spray.json.DefaultJsonProtocol.*
+import util.Logging
 
 import java.time.Instant
 import java.util.UUID
@@ -17,25 +19,27 @@ import scala.concurrent.ExecutionContext.Implicits.global
 given roomFormat: RootJsonFormat[Map[String, String]] = mapFormat[String, String]
 given roomListFormat: RootJsonFormat[List[Map[String, String]]] = listFormat(roomFormat)
 
-object MongoService:
+object MongoService extends Logging:
+  private val logger = LogManager.getLogger(getClass)
+
   def insert(json: String, room: String): Unit =
     try
       val doc = Document.parse(json)
       if !doc.containsKey("text") then
-        println("[Mongo] Skipping insert: missing 'text'")
+        logger.warn("Skipping insert: missing 'text' field")
       else if !doc.containsKey("author") then
-        println("[Mongo] Skipping insert: missing 'author'")
+        logger.warn("Skipping insert: missing 'author' field")
       else if !doc.containsKey("room") then
-        println("[Mongo] Skipping insert: missing 'room'")
+        logger.warn("Skipping insert: missing 'room' field")
       else
         MongoClientProvider.messages.insertOne(doc)
         val author = doc.get("author").asInstanceOf[org.bson.Document]
         val nickname = author.getString("nickname")
         val roomName = doc.getString("room")
-        println(s"[Mongo] Inserted message from $nickname in room $roomName")
+        logger.info(s"Inserted message from '$nickname' in room '$roomName'")
     catch
       case e: Exception =>
-        println(s"[Mongo insert error] ${e.getMessage}")
+        logger.error(s"Mongo insert error: ${e.getMessage}", e)
 
   def loadHistory(room: String, offset: Int = 0, limit: Int = 10): List[String] =
     MongoClientProvider.messages
@@ -76,10 +80,11 @@ object MongoService:
       .append("isPrivate", false)
 
     MongoClientProvider.rooms.insertOne(doc)
+    logger.info(s"Room '$name' created by user $creatorId")
     true
   }.recover {
     case e =>
-      println(s"[Mongo] Room creation error: ${e.getMessage}")
+      logger.error(s"Room creation error: ${e.getMessage}", e)
       false
   }
 
@@ -88,14 +93,17 @@ object MongoService:
     val roomOpt = Option(MongoClientProvider.rooms.find(filter).first())
 
     roomOpt.exists { room =>
-      Option(room.getString("creatorId")).contains(requesterId) && {
+      val isOwner = Option(room.getString("creatorId")).contains(requesterId)
+      if isOwner then
         MongoClientProvider.rooms.deleteOne(filter)
+        logger.info(s"Room $roomId deleted by user $requesterId")
         true
-      }
+      else
+        logger.warn(s"Unauthorized delete attempt: user $requesterId tried to delete room $roomId")
+        false
     }
   }.recover {
     case e =>
-      println(s"[Mongo] Room deletion error: ${e.getMessage}")
+      logger.error(s"Room deletion error: ${e.getMessage}", e)
       false
   }
-
